New Badge message fix
Turns out that the issue was that the badge you input is being logged as a string and the data on the sheet was an integer. I set it to text and it started working😑

I'll have to make a fix to have it recognize text or int. that way if the data refreashes it won't break anything